Prevent adding tags from breaking with divider
The frontmatter uses markdown dividers (---) to signify the
start and end. The previous search would search for the last
instance of --- in the file, which broke when using dividers
as part of the zettel content. This probably will break when
using a endmatter, but I think that case is less common.

I should probably research more robust ways to identify
front and end matters to more reliably handle this.