Adds floating window to show zettel titles
On neovim, adds a temporary floating window above zettel links with the
parsed title of the zettel while the cursor is hovering on think link.

Additionally, adds support for multiple links on the same line having
their titles displayed in the virtual text (at the end of the line).

This partially fixes #7, however I would like to replace this with just
virtual text once overlay mode has full support for virtual volumns to
handle length differences.